+++ backend/include/routing/profile/preference_index.h
@@ -2,6 +2,8 @@
 #define ROUTING_PROFILE_PREFERENCE_INDEX_H
 
 #include "routing/edges/basic_edge.h"
+#include "routing/types.h"
+
 #include "routing/database/database_helper.h"
 
 #include <string>
@@ -16,15 +18,15 @@ class PreferenceIndex{
 public:
     virtual void Load(database::DatabaseHelper& d, const std::string& index_table) = 0;
 
-    virtual void Create(database::DatabaseHelper& d, const std::vector<std::pair<unsigned_id_type, double>>& index_values, const std::string& index_table) const = 0;
+    virtual void Create(database::DatabaseHelper& d, const std::vector<std::pair<unsigned_id_type, float>>& index_values, const std::string& index_table) const = 0;
 
-    virtual double Get(unsigned_id_type uid) const = 0;
+    virtual float Get(unsigned_id_type uid) const = 0;
 
-    virtual double GetOriginal(unsigned_id_type uid) const = 0;
+    virtual float GetOriginal(unsigned_id_type uid) const = 0;
 
     virtual const std::string& GetName() const = 0;
 
-    double GetInverted(unsigned_id_type uid) const {
+    float GetInverted(unsigned_id_type uid) const {
         return kScaleMax - Get(uid);
     }
 
@@ -32,7 +34,7 @@ protected:
     /**
      * Denotes the maximum value of any value of all data indices.
      */
-    static inline const double kScaleMax = 1;
+    static inline const float kScaleMax = 1;
 
 private:
     virtual void Normalize() = 0;
